App implements static routing; home covers homepage rendering
